ACTIN-1050: move patient record json serialization out of molecular/datamodel/orange
