+++ common/src/main/kotlin/com/hartwig/actin/PatientRecordJson.kt
@@ -1,8 +1,7 @@
-package com.hartwig.actin.molecular.datamodel.orange
+package com.hartwig.actin
 
 import com.google.gson.GsonBuilder
 import com.google.gson.reflect.TypeToken
-import com.hartwig.actin.PatientRecord
 import com.hartwig.actin.clinical.datamodel.treatment.Treatment
 import com.hartwig.actin.molecular.datamodel.MolecularHistory
 import com.hartwig.actin.util.Paths
@@ -10,13 +9,13 @@ import com.hartwig.actin.util.json.GsonLocalDateAdapter
 import com.hartwig.actin.util.json.GsonLocalDateTimeAdapter
 import com.hartwig.actin.util.json.MolecularHistoryAdapter
 import com.hartwig.actin.util.json.TreatmentAdapter
+import org.apache.logging.log4j.LogManager
 import java.io.BufferedWriter
 import java.io.File
 import java.io.FileWriter
 import java.nio.file.Files
 import java.time.LocalDate
 import java.time.LocalDateTime
-import org.apache.logging.log4j.LogManager
 
 object PatientRecordJson {
     private val LOGGER = LogManager.getLogger(PatientRecordJson::class.java)
