+++ common/src/test/kotlin/com/hartwig/actin/PatientRecordJsonTest.kt
@@ -1,12 +1,12 @@
 package com.hartwig.actin
 
+import com.hartwig.actin.PatientRecordJson.fromJson
+import com.hartwig.actin.PatientRecordJson.toJson
 import com.hartwig.actin.clinical.datamodel.LabValue
 import com.hartwig.actin.clinical.interpretation.LabMeasurement
-import com.hartwig.actin.molecular.datamodel.orange.PatientRecordJson.fromJson
-import com.hartwig.actin.molecular.datamodel.orange.PatientRecordJson.toJson
-import java.time.LocalDate
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
+import java.time.LocalDate
 
 class PatientRecordJsonTest {
