--- common/src/test/resources/molecular/minimal.molecular.json
@@ -1,50 +1,0 @@
-{
-  "patientId": "irrelevant",
-  "sampleId": "irrelevant",
-  "type": "WHOLE_GENOME",
-  "refGenomeVersion": "V37",
-  "date": null,
-  "evidenceSource": "kb",
-  "externalTrialSource": "trial kb",
-  "containsTumorCells": false,
-  "hasSufficientQuality": false,
-  "hasSufficientQualityAndPurity": false,
-  "characteristics": {
-    "purity": null,
-    "ploidy": null,
-    "predictedTumorOrigin": null,
-    "isMicrosatelliteUnstable": null,
-    "microsatelliteEvidence": null,
-    "homologousRepairScore": null,
-    "isHomologousRepairDeficient": null,
-    "homologousRepairEvidence": null,
-    "tumorMutationalBurden": null,
-    "hasHighTumorMutationalBurden": null,
-    "tumorMutationalBurdenEvidence": null,
-    "tumorMutationalLoad": null,
-    "hasHighTumorMutationalLoad": null,
-    "tumorMutationalLoadEvidence": null
-  },
-  "drivers": {
-    "variants": [
-    ],
-    "copyNumbers": [
-    ],
-    "homozygousDisruptions": [
-    ],
-    "disruptions": [
-    ],
-    "fusions": [
-    ],
-    "viruses": [
-    ]
-  },
-  "immunology": {
-    "isReliable": false,
-    "hlaAlleles": [
-    ]
-  },
-  "pharmaco": [
-  ]
-}
-
